Moved some things from the UpgradeScreen script to their own scripts
