+++ Catpocalypse/Assets/Scripts/FortificationUpgrades.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7f3281cb25d5344999b6bf12d450f6f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
