+++ Catpocalypse/Assets/Scripts/Towers/Upgrades/NonAllergicUpgrades.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: abfc8be827e263d4083b5164d803c4af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
