+++ Catpocalypse/Assets/Scripts/Towers/Upgrades/ScratchingPostUpgrades.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd155cac9239d194396ddd901aaa4b76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
