+++ Catpocalypse/Assets/Scripts/Towers/Upgrades/StringWaverUpgrades.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a3afe60706566640bd01c7431951553
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
