+++ Catpocalypse/Assets/Scripts/Towers/Upgrades/YarnThrowerUpgrades.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbd2fe1e5b569a2478a37e142c18e486
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
