+++ Catpocalypse/Assets/Scripts/UpgradePanelManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31dd742c23d0f24408d2ed15a7bcf020
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
